% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raster.R
\name{find_element}
\alias{find_element}
\title{find attributes with units}
\usage{
find_element(attributes, element)
}
\description{
\code{find_element} is a helper function to
\code{create_raster} that checks for the presence of a particular element
in a list. In this case, we use \code{find_element} to determine if any of
the raster attributes metadata have units.
}
