% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_list_crs.R
\name{list_crs}
\alias{list_crs}
\title{Generate a list of EML-compliant coordinate reference system (CRS)}
\usage{
list_crs()
}
\value{
A list of EML-compliant coordinate reference systems (CRSs)
}
\description{
list_crs is a helper function that provides convenient access
to a list of EML-compliant coordinate references systems (CRSs).
}
